fix : SellBidService -> SellBidProvider 로 renaming
