+++ src/main/java/bc1/gream/domain/sell/controller/SellBidController.java
@@ -5,7 +5,7 @@ import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.sell.dto.request.SellBidRequestDto;
 import bc1.gream.domain.sell.dto.response.SellBidResponseDto;
 import bc1.gream.domain.sell.dto.response.SellCancelBidResponseDto;
-import bc1.gream.domain.sell.service.SellBidService;
+import bc1.gream.domain.sell.provider.SellBidProvider;
 import bc1.gream.global.common.RestResponse;
 import bc1.gream.global.security.UserDetailsImpl;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -25,7 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 @SecurityRequirement(name = "Bearer Authentication")
 public class SellBidController {
 
-    private final SellBidService sellBidService;
+    private final SellBidProvider sellBidProvider;
     private final ProductValidator productValidator;
 
     /**
@@ -38,7 +38,7 @@ public class SellBidController {
         @PathVariable Long productId
     ) {
         Product product = productValidator.validateBy(productId);
-        SellBidResponseDto responseDto = sellBidService.createSellBid(userDetails.getUser(), requestDto, product);
+        SellBidResponseDto responseDto = sellBidProvider.createSellBid(userDetails.getUser(), requestDto, product);
         return RestResponse.success(responseDto);
     }
 
@@ -50,7 +50,7 @@ public class SellBidController {
         @AuthenticationPrincipal UserDetailsImpl userDetails,
         @PathVariable Long sellId
     ) {
-        SellCancelBidResponseDto responseDto = sellBidService.sellCancelBid(userDetails.getUser(), sellId);
+        SellCancelBidResponseDto responseDto = sellBidProvider.sellCancelBid(userDetails.getUser(), sellId);
         return RestResponse.success(responseDto);
     }
 }
