+++ src/main/java/bc1/gream/domain/sell/provider/SellBidProvider.java
@@ -1,6 +1,6 @@
-package bc1.gream.domain.sell.service;
+package bc1.gream.domain.sell.provider;
 
-import bc1.gream.domain.gifticon.service.GifticonService;
+import bc1.gream.domain.gifticon.service.GifticonCommandService;
 import bc1.gream.domain.order.entity.Gifticon;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.sell.dto.request.SellBidRequestDto;
@@ -9,6 +9,7 @@ import bc1.gream.domain.sell.dto.response.SellCancelBidResponseDto;
 import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.sell.mapper.SellMapper;
 import bc1.gream.domain.sell.repository.SellRepository;
+import bc1.gream.domain.sell.service.SellService;
 import bc1.gream.domain.sell.service.helper.deadline.Deadline;
 import bc1.gream.domain.sell.service.helper.deadline.DeadlineCalculator;
 import bc1.gream.domain.user.entity.User;
@@ -20,15 +21,15 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class SellBidService {
+public class SellBidProvider {
 
     private final SellRepository sellRepository;
-    private final GifticonService gifticonService;
+    private final GifticonCommandService gifticonCommandService;
     private final SellService sellService;
 
     public SellBidResponseDto createSellBid(User seller, SellBidRequestDto requestDto, Product product) {
         // 기프티콘 생성, 저장
-        Gifticon gifticon = gifticonService.saveGifticon(requestDto.gifticonUrl(), null);
+        Gifticon gifticon = gifticonCommandService.saveGifticon(requestDto.gifticonUrl(), null);
         // 마감기한 지정 : LocalTime.Max :: 23시 59분 59초
         Integer period = Deadline.getPeriod(requestDto.period());
         LocalDateTime deadlineAt = DeadlineCalculator.calculateDeadlineBy(LocalDate.now(), LocalTime.MAX,
@@ -51,7 +52,7 @@ public class SellBidService {
 
     public SellCancelBidResponseDto sellCancelBid(User seller, Long sellId) {
         Sell deletedSell = sellService.deleteSellByIdAndUser(sellId, seller);
-        gifticonService.delete(deletedSell.getGifticon());
+        gifticonCommandService.delete(deletedSell.getGifticon());
 
         return new SellCancelBidResponseDto(sellId);
     }
